% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{calc_pois_pval}
\alias{calc_pois_pval}
\title{Calculate p-values based on Poisson model for each window}
\usage{
calc_pois_pval(ifs, cpois = FALSE, threshold = 1e-05)
}
\arguments{
\item{cpois}{A logical value indicating whether to calculate p-values based
on continuous Poisson model.}
}
\description{
Calculate p-values based on Poisson model for each window
}
