% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{call_hotspot}
\alias{call_hotspot}
\title{Call hotspots using FDR only}
\usage{
call_hotspot(
  ifs,
  fdr_cutoff = 0.2,
  merge_distance = 200,
  pval_cutoff = 1e-05,
  local_pval_cutoff = 1e-05,
  method = c("pois", "nb")
)
}
\description{
Call hotspots using FDR only
}
