% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_binaries}
\alias{check_binaries}
\title{Check if required binaries exist in PATH}
\usage{
check_binaries(binaries = "tabix", verbose = FALSE)
}
\arguments{
\item{binaries}{Character vector indicating what binaries to check. Default
is tabix.}

\item{verbose}{Logical value indicating whether to output detailed messages.
Default is FALSE.}
}
\value{
Logical value. TRUE if all required binaries can be found in PATH.
}
\description{
Some of cragr's functionalities require certain third-party binaries to exist
in PATH. For example, selectively loading cfDNA fragments from a certain
region under the hood invokes tabix to complete the task. This function checks
whether certain binaries exist.
}
