% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combat}
\alias{combat}
\title{Batch effect correction based on ComBat algorithm}
\usage{
combat(ifs, batch)
}
\arguments{
\item{ifs}{a data frame containing uncorrected IFS scores}

\item{batch}{a factor vector for each sample's batch information}
}
\description{
Batch effect correction based on ComBat algorithm
}
