% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{.build_cpois_lookup}
\alias{.build_cpois_lookup}
\title{Build a lookup table-based function for calculating continuous Poisson distribution CDF}
\usage{
.build_cpois_lookup(lambda, q_max, incr = 0.001)
}
\description{
Direct calculate it using \code{.build_cpois} can be expensive. An alternative is building a lookup table
}
