% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{.build_exclude_region}
\alias{.build_exclude_region}
\title{Construct the overall excluded region}
\usage{
.build_exclude_region(chrom = NULL, blacklist_region = NULL)
}
\arguments{
\item{chrom}{A character vector. If not \code{NULL}, only regions in these
chromosomes will be considered.}

\item{blacklist_region}{Can be a \code{data.table} object, a list of \code{data.table}
object, or a character vector of the data file paths.}

\item{mappability_region}{A \code{data.table} object of mappability scores, or a
character vector of the mappability score file path.}

\item{mappability_threshold}{The mappability score threshold, below which the
region is considered as low-mappability and excluded from analysis.}
}
\description{
Construct the overall excluded region
}
