% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{.exclude_region}
\alias{.exclude_region}
\title{Exclude certain regions from the dataset, using \code{bedtools}}
\usage{
.exclude_region(
  dt,
  region,
  check_sort = FALSE,
  chrom_sizes = system.file("extdata", "human_g1k_v37.chrom.sizes", package = "cragr")
)
}
\arguments{
\item{dt}{A \code{data.table} object containing fragment data.}

\item{region}{A \code{data.table} defining excluded regions.}

\item{check_sort}{\code{region} may not be correctly sorted. If \code{TRUE}, will sort
\code{region} by both chromosomes and positions, based on \code{chrom_sizes}}

\item{chrom_sizes}{Path to the chromosome size file.}
}
\value{
A filtered \code{data.table} object.
}
\description{
Exclude certain regions from the dataset, using \code{bedtools}
}
