% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.filter_by_chrom}
\alias{.filter_by_chrom}
\title{Filter the dataset by chromosomes}
\usage{
.filter_by_chrom(dt, chrom, negate = FALSE)
}
\arguments{
\item{dt}{A \code{data.table} object.}

\item{chrom}{Character vector containing chromosome names.}

\item{negate}{Indicate whether to include (\code{FALSE}) or exclude (\code{TRUE}) these
fragments.}
}
\value{
A filtered \code{data.table} object.
}
\description{
Filter the dataset by chromosomes
}
