% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.filter_by_granges}
\alias{.filter_by_granges}
\title{Filter the dataset by genomic ranges}
\usage{
.filter_by_granges(dt, granges, negate = FALSE)
}
\arguments{
\item{dt}{A \code{data.table} object.}

\item{granges}{A \code{GenomicRanges::GRanges} object, or a compatible chracter vector.}

\item{negate}{Indicate whether to include (\code{FALSE}) or exclude (\code{TRUE}) these
fragments.}
}
\value{
A filtered \code{data.table} object.
}
\description{
Filter the dataset by genomic ranges
}
