% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.filter_by_regions}
\alias{.filter_by_regions}
\title{Filter the dataset by regions}
\usage{
.filter_by_regions(dt, region, negate = FALSE)
}
\arguments{
\item{dt}{A \code{data.table} object.}

\item{region}{Can be either \code{GenomicRanges::GRanges} or character vector}

\item{negate}{Indicate whether to include (\code{FALSE}) or exclude (\code{TRUE}) these
fragments.}
}
\value{
A filtered \code{data.table} object.
}
\description{
Filter the dataset by regions
}
