% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{.gc_correct}
\alias{.gc_correct}
\title{Perform GC-correction for IFS scores}
\usage{
.gc_correct(ifs, gc, span = 0.75)
}
\arguments{
\item{ifs}{A \code{data.table} of original IFS scores.
in-place.}

\item{gc}{A \code{data.table} of GC contents.}

\item{span}{\code{span} used in LOESS fit.}
}
\value{
A \code{data.table} containing GC-corrected IFS scores/
}
\description{
Perform GC-correction for IFS scores
}
