% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.load_tabix}
\alias{.load_tabix}
\title{Selectively load fragments using tabix}
\usage{
.load_tabix(file_path, index_path = paste0(file_path, ".tbi"), regions = NULL)
}
\arguments{
\item{file_path}{Path to the tsv file.}

\item{index_path}{Path to the .tbi index file.}

\item{regions}{Which genomic region to load. Can be either a character vector
following the UCSC convention, e.g. 21:1-9411192, or a
GenomicRanges::GRanges object.}
}
\value{
A data.table object
}
\description{
Selectively load fragments using tabix
}
