% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{gc_correct}
\alias{gc_correct}
\title{Perform GC-correction for IFS scores}
\usage{
gc_correct(ifs, span = 0.75, max_training_dataset = 2000000L)
}
\arguments{
\item{span}{\code{span} used in LOESS fit.}

\item{max_training_dataset}{LOESS can be slow for a large dataset. If it
contains more data points than this parameter, we will sample a subset and
do the training.}
}
\description{
Perform GC-correction for IFS scores
}
