% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{ifs_score}
\alias{ifs_score}
\title{Calculate IFS scores}
\usage{
ifs_score(
  fragment_data,
  chrom = NULL,
  window_size = 200L,
  step_size = 20L,
  gc_correct = TRUE,
  blacklist_region = NULL,
  high_mappability_region = NULL,
  min_mapq = 30L,
  min_fraglen = 50L,
  max_fraglen = 1000L,
  exclude_soft_clipping = FALSE
)
}
\arguments{
\item{fragment_data}{A \code{GRanges} data frame containing cfDNA fragment data.}

\item{chrom}{Calculate IFS scores for certain chromosomes. If \code{NULL}, all
chromosomes in \code{fragment_data} will be used.}

\item{window_size}{Width of the sliding window. Must be multiples of \code{step_size}.}

\item{step_size}{Incremental steps of the sliding window.}

\item{gc_correct}{Logical value. Whether to perform GC correction. If \code{TRUE},
the corresponding \code{BSgenome} should be present. Currently, only support
\code{BSgenome.Hsapiens.1000genomes.hs37d5}.}

\item{blacklist_region}{Those fragments whose midpoint fall within any of the
excluded regions will not be used in the analysis. \code{blacklist_region} can
be either a character vector that contains the names of the files defining
the regions (should be in BED format), a \code{data.table} object, or a list of
\code{data.table} objects.}

\item{high_mappability_region}{A \code{data.table} object of mappability scores, or a
character vector of the mappability score file path.}

\item{min_mapq}{Exclude fragments whose \code{MAPQ} scores are smaller than \code{min_mapq}.}

\item{min_fraglen}{Exclude fragments shorter than \code{min_fraglen}.}

\item{max_fraglen}{Exclude fragments longer than \code{max_fraglen}.}

\item{exclude_soft_clipping}{For fragments with leading soft clipping, their
inferred lengths are less reliable. If \code{TRUE}, such fragments will be
excluded from the analysis. \code{cigar1} and \code{cigar2} should be present in the
dataset.}

\item{mappability_threshold}{The mappability score threshold, below which the
region is considered as low-mappability and excluded from analysis.}
}
\description{
Calculate IFS scores
}
