% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jaccard_index}
\alias{jaccard_index}
\title{Compare two hotspot files and return the Jaccard index}
\usage{
jaccard_index(hotspot1, hotspot2, max_distance = 200L)
}
\description{
Calculate Jaccard index using merged hotspots
}
