% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{postprocess_ifs}
\alias{postprocess_ifs}
\title{Postprocess raw IFS scores}
\usage{
postprocess_ifs(ifs, chrom, high_mappability_region, gc_correct)
}
\description{
\enumerate{
\item Mappability filter
\item GC-correction
\item Calculate z-scores
}
}
\details{
This is counter-intuitive, but use much less memory and SU
}
