% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_fragments}
\alias{read_fragments}
\title{Read fragment BED file}
\usage{
read_fragments(file_path, range = NULL, genome = NULL, verbose = FALSE)
}
\arguments{
\item{genome}{Should be GRCh37 or GRCh38}
}
\description{
Read fragment BED file
}
