% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{slide_window}
\alias{slide_window}
\title{Perform a sliding window over raw IFS scores}
\usage{
slide_window(ifs, window_size, step_size)
}
\arguments{
\item{window_size}{Width of the sliding window. Must be multiples of \code{step_size}.}

\item{step_size}{Incremental steps of the sliding window.}
}
\description{
Raw IFS scores are calculated for each \code{step_size} bin. This function
performs a sliding window with the width of \code{window_size}. Each window
contains a whole number of \code{step_size} bins.
}
