% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_bed}
\alias{write_bed}
\title{Write data frame to disk in BED format}
\usage{
write_bed(dt, file_path, col_names = TRUE, create_index = FALSE, ...)
}
\arguments{
\item{dt}{Data frame to write.}

\item{file_path}{Output file name.}

\item{col_names}{A logical value indicating whether to output colum names.}

\item{create_index}{A logical value indicating whether to create an index
file. Ignored if the output is not gzipped.}
}
\description{
If the target file indicates gzip-compression, will invoke bgzip rather than
standard gzip. Another feature is, if the output file should contain header
lines, they will start with \verb{#}, as specified in BED format specification.
}
