% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{call_hotspot}
\alias{call_hotspot}
\title{Call hotspots based on pvalues (and merge)}
\usage{
call_hotspot(
  ifs,
  use_cpois = FALSE,
  fdr_cutoff = 0.01,
  pval_cutoff = 1e-05,
  local_pval_cutoff = 1e-05,
  merge_distance = 200
)
}
\description{
Call hotspots based on pvalues (and merge)
}
