% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{gc_correct}
\alias{gc_correct}
\title{Perform GC-correction for IFS scores}
\usage{
gc_correct(
  ifs,
  span = 0.75,
  max_training_dataset = 1000000L,
  method = "standard",
  return_model = FALSE,
  ...
)
}
\arguments{
\item{span}{\code{span} used in LOESS fit.}

\item{max_training_dataset}{LOESS can be slow for a large dataset. If it
contains more data points than this parameter, we will sample a subset and
do the training.}

\item{method}{If \code{standard}, the built-in \code{loess} function will be used for
the calculation. If \code{caret}, will use \code{gamLOESS}}
}
\description{
Perform GC-correction for IFS scores
}
