% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{ifs_score}
\alias{ifs_score}
\title{Calculate IFS scores}
\usage{
ifs_score(
  fragment_data,
  chrom = NULL,
  excluded_regions = NULL,
  chrom_sizes = system.file("extdata", "human_g1k_v37.chrom.sizes", package = "cragr"),
  min_mapq = 30,
  min_fraglen = 50,
  max_fraglen = 1000,
  exclude_soft_clipping = FALSE
)
}
\arguments{
\item{fragment_data}{A \code{data.table} containing cfDNA fragment data.}

\item{chrom}{Calculate IFS scores for certain chromosomes. If \code{NULL}, all
chromosomes in \code{fragment_data} will be used.}

\item{excluded_regions}{Those fragments whose midpoint fall within any of the
excluded regions will not be used in the analysis. \code{excluded_regions} can
be either a character vector that contains the names of the files defining
the regions (should be in BED format), or a list of \code{data.table} objects.}

\item{chrom_sizes}{Path to the chromosome size file.}

\item{min_mapq}{Exclude fragments whose \code{MAPQ} scores are smaller than \code{min_mapq}.}

\item{min_fraglen}{Exclude fragments shorter than \code{min_fraglen}.}

\item{max_fraglen}{Exclude fragments longer than \code{max_fraglen}.}

\item{exclude_soft_clipping}{For fragments with leading soft clipping, their
inferred lengths are less reliable. If \code{TRUE}, such fragments will be
excluded from the analysis.}
}
\description{
Calculate IFS scores
}
