% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs.R
\name{postprocess_ifs}
\alias{postprocess_ifs}
\title{Post-process raw IFS scores}
\usage{
postprocess_ifs(ifs, gc_correct = TRUE)
}
\arguments{
\item{ifs}{Result returned by \code{ifs_score()}.}

\item{gc_correct}{A logical value indicating whether to perform
GC-correction.}
}
\description{
From result from \code{ifs_score()}:
\enumerate{
\item Perform GC-correction
\item Calculate z-scores
}
}
